+++ src/Resources/app/storefront/src/page/swag-paypal.express-checkout.js
@@ -103,7 +103,11 @@ export default class SwagPayPalExpressCheckoutButton extends SwagPaypalAbstractB
 
     createButton() {
         this.createScript((paypal) => {
-            this.renderButton(paypal);
+            try {
+                this.renderButton(paypal);
+            } catch (error) {
+                this.handleError(this.SCRIPT_ERROR, true, error);
+            }
         });
     }